Move progress bar widget out of kartapi and into dockwidget
This keeps the UI concerns in the GUI component and removes the UI
dependancy in kartapi. An output_handler hook is added to
Repository.clone to enable progress reporting.
